Completed the drawing of the badge.
Before applying scaleToFit in the BadgeView zstack view, the geometry reader presents all the symbols way to close to the border.
Applying that scaleToFit allows you to contain all the drawing to the size of the background (it seems) so it ends up reducing the canvas of the geometry reader. Interesting stuff.