Not happy Jen
For some reason the project does not compile/run as indicated by Apple. Had to do a few changes and the NavigationLink does not trigger the navigation for a given element in the categoryRow. I was debugging trying to find the root cause of it, but it seems it is just the level at which the navigation item is.